refactor(experiments): extract the configure store's helpers and constants
Pure functions move to util/dot-experiments-configure.util.ts and the i18n
keys plus weight constants to shared/constants.ts, matching the list's
layout. Kills three duplicate '/experiments' literals and a second
TOTAL_WEIGHT along the way.
